+++ src/main/resources/db/changelog/changeset/6.create-beer-characteristic-table.sql
@@ -3,7 +3,7 @@
 create type beer_characteristic_type AS ENUM ('ABV', 'OG', 'IBU', 'EBC');
 create table beer_characteristic
 (
-beer_id integer references beer (id) on delete cascade,
+beer_id bigint references beer (id) on delete cascade,
 characteristic beer_characteristic_type not null,
 value double precision not null,
 primary key (beer_id,characteristic)
